requirements.txt
faker==22.6.0
neo4j==5.15.0
python-dotenv==1.0.1 
fastapi==0.117.1
uvicorn==0.37.0